requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.16.47
botocore==1.19.47
certifi==2020.12.5
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-adaptors==0.2.5
django-import-export==2.3.0
django-render-block==0.8.1
django-storages==1.11.1
django-templated-email==2.3.0
et-xmlfile==1.0.1
Faker==5.0.1
gunicorn==20.0.4
isort==5.5.4
jdcal==1.4.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
psycopg2-binary==2.8.6
pylint==2.6.0
pylint-django==2.4.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
toml==0.10.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
